% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codigospostales}
\alias{codigospostales}
\title{KML's con Indicadores por Codigo Postal}
\format{Una lista con SpatialPolygonDataFrames, uno por cada estado.}
\source{
\url{http://catalogo.datos.gob.mx/dataset/ubicacion-de-codigos-postales-en-mexico}
}
\usage{
codigospostales
}
\description{
Una lista de KML's manipulados en R, con indicadores preliminares, para usarse en Mapas de Google.
}
\keyword{datasets}

